# docs/web/api/openapi.yaml
openapi: 3.1.1
info:
  description: API documentation for universe products service
  title: universe products API
  version: 0.1.0
servers:
- url: http://localhost:8001
paths:
  /products:
    get:
      description: "Endpoint to retrieve a list of products with filtering, sorting\
        \ and pagination."
      parameters:
      - description: Max number of items to return
        explode: true
        in: query
        name: "page[limit]"
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - description: Number of items to skip
        explode: true
        in: query
        name: "page[offset]"
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      - description: Filter by product name (partial match)
        explode: true
        in: query
        name: "filter[name]"
        required: false
        schema:
          type: string
        style: form
      - description: Filter products with price greater than or equal to this value
        explode: true
        in: query
        name: "filter[price][gte]"
        required: false
        schema:
          format: double
          minimum: 0
          type: number
        style: form
      - description: Filter products with price less than or equal to this value
        explode: true
        in: query
        name: "filter[price][lte]"
        required: false
        schema:
          format: double
          minimum: 0
          type: number
        style: form
      - description: Filter products created at or after this timestamp (RFC3339)
        example: 2024-01-01T00:00:00Z
        explode: true
        in: query
        name: "filter[created_at][gte]"
        required: false
        schema:
          format: date-time
          type: string
        style: form
      - description: Filter products created at or before this timestamp (RFC3339)
        example: 2024-12-31T23:59:59Z
        explode: true
        in: query
        name: "filter[created_at][lte]"
        required: false
        schema:
          format: date-time
          type: string
        style: form
      - description: |
          Sort field and direction. Prefix `-` means descending order. Example: `sort=price` (cheapest first), `sort=-price` (most expensive first)
        explode: true
        in: query
        name: sort
        required: false
        schema:
          default: -created_at
          enum:
          - price
          - -price
          - created_at
          - -created_at
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductsCollection"
          description: A list of products
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
          description: Invalid query parameters
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
          description: Internal server error
      summary: Get products
      tags:
      - products
    post:
      description: Endpoint to create a new product.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
          description: Product successfully created
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
          description: Bad request - invalid input data
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
          description: Internal server error
      summary: Create a new product
      tags:
      - products
  /products/{product_id}:
    delete:
      description: Endpoint to delete a product by its ID.
      parameters:
      - description: product ID
        explode: false
        in: path
        name: product_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: Product successfully deleted
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
          description: Bad request - invalid product ID
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
          description: Product not found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Errors"
          description: Internal server error
      summary: Delete a product
      tags:
      - products
    parameters:
    - description: product ID
      explode: false
      in: path
      name: product_id
      required: true
      schema:
        format: uuid
        type: string
      style: simple
components:
  schemas:
    Errors:
      description: Standard JSON:API error
      example:
        errors:
        - detail: Request body was expected
          title: Bad Request
          status: 400
        - detail: Request body was expected
          title: Bad Request
          status: 400
      properties:
        errors:
          description: Non empty array of errors occurred during request processing
          items:
            $ref: "#/components/schemas/Errors_errors_inner"
          type: array
      required:
      - errors
    Product:
      example:
        data:
          attributes:
            price: 0.8008281904610115
            name: name
            created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          type: product
      properties:
        data:
          $ref: "#/components/schemas/ProductData"
      required:
      - data
    ProductsCollection:
      example:
        data:
        - attributes:
            price: 0.8008281904610115
            name: name
            created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          type: product
        - attributes:
            price: 0.8008281904610115
            name: name
            created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          type: product
        links:
          next: https://openapi-generator.tech
          last: https://openapi-generator.tech
          prev: https://openapi-generator.tech
          self: https://openapi-generator.tech
          first: https://openapi-generator.tech
      properties:
        data:
          items:
            $ref: "#/components/schemas/ProductData"
          type: array
        links:
          $ref: "#/components/schemas/PaginationData"
      required:
      - data
      - links
    PaginationData:
      example:
        next: https://openapi-generator.tech
        last: https://openapi-generator.tech
        prev: https://openapi-generator.tech
        self: https://openapi-generator.tech
        first: https://openapi-generator.tech
      properties:
        self:
          description: link to the current page
          format: uri
          type: string
        first:
          description: link to the first page
          format: uri
          type: string
        last:
          description: link to the last page
          format: uri
          type: string
        prev:
          description: link to the previous page
          format: uri
          type: string
        next:
          description: link to the next page
          format: uri
          type: string
      required:
      - self
    ProductData:
      example:
        attributes:
          price: 0.8008281904610115
          name: name
          created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        type: product
      properties:
        id:
          description: product ID
          format: uuid
          type: string
        type:
          enum:
          - product
          type: string
        attributes:
          $ref: "#/components/schemas/ProductData_attributes"
      required:
      - attributes
      - id
      - type
    CreateProduct:
      example:
        data:
          attributes:
            price: 0.8008281904610115
            name: name
          type: product
      properties:
        data:
          $ref: "#/components/schemas/CreateProduct_data"
      required:
      - data
    Errors_errors_inner:
      example:
        detail: Request body was expected
        title: Bad Request
        status: 400
      properties:
        title:
          description: "Title is a short, human-readable summary of the problem"
          example: Bad Request
          type: string
        detail:
          description: Detail is a human-readable explanation specific to this occurrence
            of the problem
          example: Request body was expected
          type: string
        status:
          description: Status is the HTTP status code applicable to this problem
          enum:
          - 400
          - 401
          - 403
          - 404
          - 409
          - 500
          example: 400
          type: integer
      required:
      - status
      - title
    ProductData_attributes:
      example:
        price: 0.8008281904610115
        name: name
        created_at: 2000-01-23T04:56:07.000+00:00
      properties:
        name:
          description: The name of the product
          type: string
        price:
          description: The price of the product
          format: double
          type: number
        created_at:
          description: The date and time when the product was created
          format: date-time
          type: string
      required:
      - created_at
      - name
      - price
    CreateProduct_data_attributes:
      example:
        price: 0.8008281904610115
        name: name
      properties:
        name:
          description: The name of the product
          type: string
        price:
          description: The price of the product
          format: double
          type: number
      required:
      - name
      - price
    CreateProduct_data:
      example:
        attributes:
          price: 0.8008281904610115
          name: name
        type: product
      properties:
        type:
          enum:
          - product
          type: string
        attributes:
          $ref: "#/components/schemas/CreateProduct_data_attributes"
      required:
      - attributes
      - type